+++ types/affinda.d.ts
@@ -1,9 +1,9 @@
-import { AccessToken } from '@azure/identity';
+import { AccessToken } from '@azure/core-auth';
 import * as coreAuth from '@azure/core-auth';
 import * as coreClient from '@azure/core-client';
 import * as coreRestPipeline from '@azure/core-rest-pipeline';
-import { GetTokenOptions } from '@azure/identity';
-import { TokenCredential } from '@azure/identity';
+import { GetTokenOptions } from '@azure/core-auth';
+import { TokenCredential } from '@azure/core-auth';
 
 export declare interface Accreditation {
     education?: string;
@@ -901,10 +901,6 @@ export declare interface AffindaAPIGetAllDataPointsOptionalParams extends coreCl
     description?: string;
     /** Filter by annotation content type, e.g. text, integer, float, date, etc. */
     annotationContentType?: string;
-    /** Whether to show child data points at the top level. <br /> By default child data points are shown nested inside their parent so they are excluded from the top level. */
-    includeChild?: boolean;
-    /** Filter by specific identifiers. */
-    identifier?: string[];
 }
 
 /** Contains response data for the getAllDataPoints operation. */
@@ -1757,12 +1753,6 @@ export declare type CurrencyCodeAnnotation = Annotation & {
     parsed?: DataPointChoice;
 };
 
-export declare interface CustomFieldConfig {
-    /** Data point identifier. */
-    dataPoint: string;
-    weight: number;
-}
-
 export declare interface DataPoint {
     /** Uniquely identify a data point. */
     identifier: string;
@@ -1856,7 +1846,7 @@ export declare type DateFormatPreference = string;
  */
 export declare type DateRange = string;
 
-export declare interface Document {
+declare interface Document_2 {
     /** Polymorphic discriminator, which specifies the different types this object can be */
     extractor: "resume" | "invoice" | "job-description";
     /** Dictionary of <any> */
@@ -1866,6 +1856,7 @@ export declare interface Document {
     meta: DocumentMeta;
     error?: DocumentError;
 }
+export { Document_2 as Document }
 
 export declare interface DocumentCreate {
     /** File as binary data blob. Supported formats: PDF, DOC, DOCX, TXT, RTF, HTML, PNG, JPG */
@@ -2016,7 +2007,7 @@ export declare interface DocumentSplitPage {
  */
 export declare type DocumentState = string;
 
-export declare type DocumentUnion = Document | Resume | Invoice | JobDescription;
+export declare type DocumentUnion = Document_2 | Resume | Invoice | JobDescription;
 
 export declare interface DocumentUpdate {
     /** Uniquely identify a collection. */
@@ -2037,7 +2028,7 @@ export declare interface Education {
     organization?: string;
     accreditation?: Accreditation;
     grade?: EducationGrade;
-    location?: Location;
+    location?: Location_2;
     dates?: EducationDates;
 }
 
@@ -2254,7 +2245,7 @@ export declare interface InvitationUpdate {
     role?: OrganizationRole;
 }
 
-export declare type Invoice = Document & {
+export declare type Invoice = Document_2 & {
     /** Polymorphic discriminator, which specifies the different types this object can be */
     extractor: "invoice";
     data?: InvoiceData;
@@ -2366,7 +2357,7 @@ export declare interface InvoiceDataTablesItem {
     rows?: (RowAnnotation | null)[];
 }
 
-export declare type JobDescription = Document & {
+export declare type JobDescription = Document_2 & {
     /** Polymorphic discriminator, which specifies the different types this object can be */
     extractor: "job-description";
     data?: JobDescriptionData;
@@ -2447,7 +2438,6 @@ export declare interface JobDescriptionSearchConfig {
     actions?: SearchConfigAction[];
     /** Hide the reset/import toolbar. */
     hideToolbar?: boolean;
-    customFieldsConfig?: CustomFieldConfig[];
 }
 
 export declare interface JobDescriptionSearchDetail {
@@ -2510,7 +2500,7 @@ export declare interface JobDescriptionSearchDetailLocation {
     value?: JobDescriptionSearchDetailLocationValue;
 }
 
-export declare type JobDescriptionSearchDetailLocationValue = Location & Components1TlnsonSchemasJobdescriptionsearchdetailPropertiesLocationPropertiesValueAllof1 & {};
+export declare type JobDescriptionSearchDetailLocationValue = Location_2 & Components1TlnsonSchemasJobdescriptionsearchdetailPropertiesLocationPropertiesValueAllof1 & {};
 
 export declare interface JobDescriptionSearchDetailManagementLevel {
     level?: ManagementLevel;
@@ -2823,7 +2813,7 @@ export declare interface LanguagesSearchScoreComponent {
     score?: number;
 }
 
-export declare interface Location {
+declare interface Location_2 {
     /** NOTE: This property will not be serialized. It can only be populated by the server. */
     readonly formatted?: string;
     /** NOTE: This property will not be serialized. It can only be populated by the server. */
@@ -2851,9 +2841,10 @@ export declare interface Location {
     /** NOTE: This property will not be serialized. It can only be populated by the server. */
     readonly longitude?: number;
 }
+export { Location_2 as Location }
 
 export declare type LocationAnnotation = Annotation & {
-    parsed?: Location;
+    parsed?: Location_2;
 };
 
 export declare interface LocationSearchScoreComponent {
@@ -3227,7 +3218,7 @@ export declare interface ResthookSubscriptionUpdate {
  */
 export declare type ResthookSubscriptionVersion = string;
 
-export declare type Resume = Document & {
+export declare type Resume = Document_2 & {
     /** Polymorphic discriminator, which specifies the different types this object can be */
     extractor: "resume";
     /** A JSON-encoded string of the `ResumeData` object. */
@@ -3243,7 +3234,7 @@ export declare interface ResumeData {
     websites?: string[];
     emails?: string[];
     dateOfBirth?: string;
-    location?: Location;
+    location?: Location_2;
     objective?: string;
     /** NOTE: This property will not be serialized. It can only be populated by the server. */
     readonly languages?: string[];
@@ -3348,7 +3339,7 @@ export declare interface ResumeDataWorkExperienceItem {
     organization?: string;
     /** NOTE: This property will not be serialized. It can only be populated by the server. */
     readonly industry?: string;
-    location?: Location;
+    location?: Location_2;
     jobDescription?: string;
     dates?: ResumeDataWorkExperienceItemDates;
     /** NOTE: This property will not be serialized. It can only be populated by the server. */
@@ -3427,7 +3418,6 @@ export declare interface ResumeSearchConfig {
     actions?: SearchConfigAction[];
     /** Hide the reset/import toolbar. */
     hideToolbar?: boolean;
-    customFieldsConfig?: CustomFieldConfig[];
 }
 
 export declare interface ResumeSearchDetail {
@@ -3487,7 +3477,7 @@ export declare interface ResumeSearchDetailLocation {
     value?: ResumeSearchDetailLocationValue;
 }
 
-export declare type ResumeSearchDetailLocationValue = Location & ComponentsN9ShogSchemasResumesearchdetailPropertiesLocationPropertiesValueAllof1 & {};
+export declare type ResumeSearchDetailLocationValue = Location_2 & ComponentsN9ShogSchemasResumesearchdetailPropertiesLocationPropertiesValueAllof1 & {};
 
 export declare interface ResumeSearchDetailManagementLevel {
     level?: ManagementLevel;