+++ CHANGELOG.md
@@ -5,14 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Security
+- Remove dependency on azure/identity, bump other versions to get security fixes
 
 ## [6.4.0] - 2023-04-18
 ### Changed
 - rawText is now not nullable
 - OccupationGroupSearchResult.children is now optional
-
-### Added
-- Add custom fields to resume and job search configs
 
 ## [6.3.2] - 2023-03-29
 ### Fixed