+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [6.5.0] - 2023-04-19
 ### Security
 - Remove dependency on azure/identity, bump other versions to get security fixes
 
@@ -12,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - rawText is now not nullable
 - OccupationGroupSearchResult.children is now optional
+
+### Added
+- Add custom fields to resume and job search configs
 
 ## [6.3.2] - 2023-03-29
 ### Fixed